move: SectionHeader, ChannelCard to src/components
