+++ src/components/SectionHeader.tsx
@@ -1,20 +1,24 @@
 import styled from "styled-components";
 
-const Wrapper = styled.div`
+type dynamicColor = {
+  color: string;
+}
+
+const Wrapper = styled.div<dynamicColor>`
   width: 100%;
   display: flex;
   justify-content: space-between;
   font-family: SBAggroM;
   font-size: 1.5rem;
   padding: 1rem;
-  background: var(--purple);
+  background-color: ${props=>props.color};
   border-bottom: 1px solid var(--white);
   text-shadow: 0 2px 0 black;
 `;
 
-const SectionHeader = ({ children, title }: any) => {
+const SectionHeader = ({ children, title, color }: any) => {
   return (
-    <Wrapper>
+    <Wrapper color={color}>
       {title}
       {children}
     </Wrapper>
